renamed MovieList to PopularMovies and moved MovieDetails to PopularMovies
